+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.1.1 "Milano" (03.05.2015 - 04.05.2015)
 * XLSXML() into- function with colors
-
+* $$hashKey - remove Angular's key
 
 ### 0.1.0 (aka 0.0.52) "Venice" (02.05.2015 - 03.05.2015)
 * Added INFORMATION_SCHEMA from variable